Deduplicate tree and column filter helpers
StandardTreeItem and MultiDBTreeItem duplicated the filtered-children
row_count/child scaffolding. Extract a FilterableChildrenMixin holding
the shared surface (the _compute_visible_children hook plus row_count
and child), while each item keeps its own deliberately different cache
and O(1) child_number lookup. Also drop the pass-through
_make_column_matcher wrapper in favor of the shared make_search_matcher.
