+++ spinetoolbox/mvcmodels/minimal_tree_model.py
@@ -229,6 +229,43 @@ class TreeItem:
         raise NotImplementedError()
 
 
+class FilterableChildrenMixin:
+    """Routes Qt row access through a filtered ``visible_children`` view instead of raw ``children``.
+
+    Mixed into tree items that may hide some children behind a level/search filter. It captures only the
+    parts that are identical no matter *how* the visible list is cached: :meth:`row_count`/:meth:`child`
+    read the (subclass-cached) ``visible_children`` list, and :meth:`_compute_visible_children` is the
+    extension point that produces the filtered list (the base passes every child through).
+
+    Deliberately left to the concrete item classes, because their post-optimization strategies differ and
+    that differentiation is correctness-critical:
+
+    - ``visible_children`` and its caching/invalidation (one memoizes on a ``filter_generation`` counter,
+      the other rebuilds alongside its child map), and
+    - ``child_number`` (an O(1) lookup keyed differently in each class).
+
+    This mixin does NOT try to unify those.
+    """
+
+    def _compute_visible_children(self):
+        """Returns the children that pass the active filters. Overridden by filtering subclasses/items.
+
+        The base does no filtering, so all children are visible.
+        """
+        return self._children
+
+    def row_count(self):
+        """Overridden to count only visible children."""
+        return len(self.visible_children)
+
+    def child(self, row):
+        """Overridden to return the visible child at the given row or None if out of bounds."""
+        visible = self.visible_children
+        if 0 <= row < len(visible):
+            return visible[row]
+        return None
+
+
 class MinimalTreeModel(QAbstractItemModel):
     """Base class for all tree models."""
